refator: merge `status` and `reports` command
