+++ packages/yuudachi/src/commands/moderation/reports.ts
@@ -1,37 +1,30 @@
-import { Collection, type Message, type Snowflake } from 'discord.js';
+import { Collection, type Snowflake } from 'discord.js';
 import i18next from 'i18next';
-import type { Sql } from 'postgres';
-import { container } from 'tsyringe';
+import { lookup } from './sub/reports/lookup.js';
+import { status } from './sub/reports/status.js';
 import { type ArgsParam, Command, type InteractionParam, type LocaleParam, type CommandMethod } from '../../Command.js';
 import {
 	AUTOCOMPLETE_CHOICE_LIMIT,
 	AUTOCOMPLETE_CHOICE_NAME_LENGTH_LIMIT,
 	SNOWFLAKE_MIN_LENGTH,
 } from '../../Constants.js';
-import { formatMessageToEmbed } from '../../functions/logging/formatMessageToEmbed.js';
-import { generateReportEmbed } from '../../functions/logging/generateReportEmbed.js';
 import { ReportType } from '../../functions/reports/createReport.js';
 import { findReports } from '../../functions/reports/findReports.js';
-import { type RawReport, transformReport } from '../../functions/reports/transformReport.js';
-import type { ReportLookupCommand } from '../../interactions/index.js';
+import type { ReportUtilsCommand } from '../../interactions/index.js';
 import { logger } from '../../logger.js';
-import { kSQL } from '../../tokens.js';
 import { REPORT_KEYS } from '../../util/actionKeys.js';
-import { ellipsis, truncateEmbed } from '../../util/embed.js';
-import { generateHistory, generateUserInfo, HistoryType } from '../../util/generateHistory.js';
-import { resolveMemberAndUser } from '../../util/resolveMemberAndUser.js';
-import { resolveMessage } from '../../util/resolveMessage.js';
+import { ellipsis } from '../../util/embed.js';
 
 const OP_DELIMITER = '-' as const;
 
-export default class extends Command<typeof ReportLookupCommand> {
+export default class extends Command<typeof ReportUtilsCommand> {
 	public override async autocomplete(
 		interaction: InteractionParam<CommandMethod.Autocomplete>,
-		args: ArgsParam<typeof ReportLookupCommand>,
+		args: ArgsParam<typeof ReportUtilsCommand>,
 		locale: LocaleParam,
 	): Promise<void> {
 		try {
-			const trimmedPhrase = args.phrase.trim();
+			const trimmedPhrase = args.lookup.phrase.trim();
 			const reports = await findReports(trimmedPhrase, interaction.guildId);
 			let choices = reports.map((r) => {
 				const choiceName = `#${r.report_id} ${r.type === ReportType.Message ? '✉️' : '👤'} ${REPORT_KEYS[
@@ -124,63 +117,20 @@ export default class extends Command<typeof ReportLookupCommand> {
 
 	public override async chatInput(
 		interaction: InteractionParam,
-		args: ArgsParam<typeof ReportLookupCommand>,
+		args: ArgsParam<typeof ReportUtilsCommand>,
 		locale: LocaleParam,
 	): Promise<void> {
-		const sql = container.resolve<Sql<any>>(kSQL);
-		await interaction.deferReply({ ephemeral: args.hide ?? true });
+		// eslint-disable-next-line @typescript-eslint/no-unnecessary-condition
+		await interaction.deferReply({ ephemeral: args.lookup?.hide ?? true });
 
-		const [cmd, id] = args.phrase.split(OP_DELIMITER);
-
-		if (cmd === 'history' && id) {
-			const data = await resolveMemberAndUser(interaction.guild, id);
-
-			const embed = truncateEmbed(await generateHistory(interaction, data, locale, HistoryType.Report));
-
-			await interaction.editReply({
-				embeds: [embed],
-			});
-			return;
+		switch (Object.keys(args)[0]) {
+			case 'lookup':
+				await lookup(interaction, args.lookup, locale);
+				break;
+			case 'status':
+				await status(interaction, args.status, locale);
+				break;
+			default:
 		}
-
-		if (!isNaN(parseInt(args.phrase, 10))) {
-			const [report] = await sql<RawReport[]>`
-			select *
-			from reports
-			where guild_id = ${interaction.guildId}
-			and report_id = ${args.phrase}`;
-
-			if (!report) {
-				throw new Error(i18next.t('command.common.errors.use_autocomplete', { lng: locale }));
-			}
-
-			let message: Message | undefined = undefined;
-
-			try {
-				message = report.message_id
-					? await resolveMessage(interaction.channelId, report.guild_id, report.channel_id, report.message_id, locale)
-					: undefined;
-			} catch {}
-
-			const author = await interaction.client.users.fetch(report.author_id);
-
-			const embeds = [truncateEmbed(await generateReportEmbed(author, transformReport(report), locale, message))];
-
-			if (message) {
-				embeds.push(truncateEmbed(await formatMessageToEmbed(message as Message<true>, locale)));
-			}
-
-			if (report.type === ReportType.User) {
-				const target = await resolveMemberAndUser(interaction.guild, report.target_id);
-				embeds.push(truncateEmbed(generateUserInfo(target, locale)));
-			}
-
-			await interaction.editReply({
-				embeds,
-			});
-			return;
-		}
-
-		throw new Error(i18next.t('command.common.errors.use_autocomplete', { lng: locale }));
 	}
 }
