+++ packages/yuudachi/src/interactions/index.ts
@@ -13,7 +13,6 @@ export * from "./moderation/unban.js";
 export * from "./moderation/warn.js";
 export * from "./moderation/timeout.js";
 export * from "./moderation/clear.js";
-export * from "./moderation/status.js";
 export * from "./moderation/reports.js";
 export * from "./utility/ping.js";
 export * from "./utility/checkScam.js";
